desktop-tool/requirements.txt
attrs~=23.1
click==8.0.4
colorama~=0.4
coverage~=7.2
defusedxml~=0.7
enlighten~=1.11
fpdf2~=2.7
InquirerPy~=0.3
pillow==9.5.0
pre-commit
pyinstaller~=5.10
pytest~=7.3
ratelimit~=2.2
requests~=2.30
sanitize-filename~=1.2
selenium~=4.8
webdriver-manager==3.8.6
wakepy==0.6.0
